move solara_viz back to experimental
