+++ mesa/experimental/__init__.py
@@ -1,3 +1,4 @@
 from mesa.experimental import cell_space
+from .solara_viz import JupyterViz, SolaraViz, make_text, Slider
 
-__all__ = ["cell_space"]
+__all__ = ["cell_space", "JupyterViz", "SolaraViz", "make_text", "Slider"]
