+++ mesa/experimental/components/matplotlib.py
