--- mesa/visualization/__init__.py
@@ -1,3 +1,0 @@
-from .solara_viz import JupyterViz, Slider, SolaraViz, make_text
-
-__all__ = ["JupyterViz", "make_text", "Slider", "SolaraViz"]
